fx(refactor): - trim input strings - center modals - minor css fixes - create socket controller to handle sockets - fix resolve helper
