+++ client/src/components/Login.vue
@@ -6,7 +6,7 @@
           <div class="form-icon"><i class="fa fa-envelope"></i></div>
         </b-input-group-prepend>
         <b-form-input
-          v-model="form.email"
+          v-model.trim="form.email"
           type="email"
           required
           placeholder="Enter your email"
