+++ server/controllers/index.js
@@ -1,2 +1,3 @@
 export { default as User } from './users';
 export { default as Game } from './game';
+export { default as SocketController } from './socket';
